+++ deltafi-ui/CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.103.0] (Unreleased)
+## [0.104.0] (Unreleased)
+
+### Fixed
+
+- Fixed tooltips on trace chart
+
+## [0.103.0]
 
 ### Fixed